tests/wpt/web-platform-tests/largest-contentful-paint/image-src-change.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Largest Contentful Paint: src change triggers new entry.</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/largest-contentful-paint-helpers.js"></script>
<img src='/images/blue.png' id='image_id'/>
<script>
  async_test(function (t) {
    assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");
    let beforeLoad = performance.now();
    let firstCallback = true;
    const observer = new PerformanceObserver(
      t.step_func(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        const url = window.location.origin + (firstCallback ? '/images/blue.png' : '/images/black-rectangle.png');
        // blue.png is 133 by 106. black-rectangle.png is 100 x 50.
        const size = firstCallback ? 133 * 106 : 100 * 50;
        checkImage(entry, url, 'image_id', size, beforeLoad);
        if (firstCallback) {
          // Set the src to trigger another entry.
          const img = document.getElementById('image_id');
          img.src = '/images/black-rectangle.png';
          beforeLoad = performance.now();
          firstCallback =false;
        } else {
          t.done();
        }
      })
    );
    observer.observe({type: 'largest-contentful-paint', buffered: true});
  }, 'Largest Contentful Paint: changing src causes a new entry to be dispatched.');
</script>
</body>
